% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{creator}
\alias{creator}
\title{Creator (author) keyword search in Zotero DB}
\usage{
creator(keyword, con = NULL, zotero_db = "~/Zotero/zotero.sqlite")
}
\arguments{
\item{keyword}{Creator (author) keyword (partial match)}

\item{con}{An existing DBI connection to a Zotero SQLite DB (optional)}

\item{zotero_db}{Path to the Zotero SQLite DB (default: '~/Zotero/zotero.sqlite')}
}
\value{
A data.frame with columns: itemID, title, creators, collections, itemTypeID, dateAdded, dateModified, key
}
\description{
Creator (author) keyword search in Zotero DB
}
