% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{key}
\alias{key}
\title{Search Zotero DB collection by name}
\usage{
key(key, con = NULL, zotero_db = "~/Zotero/zotero.sqlite")
}
\arguments{
\item{key}{Zotero item key (exact match)}

\item{con}{An existing DBI connection to a Zotero SQLite DB (optional)}

\item{zotero_db}{Path to the Zotero SQLite DB (default: '~/Zotero/zotero.sqlite')}
}
\value{
A data.frame with columns: itemID, title, creators, collections, itemTypeID, dateAdded, dateModified, key
}
\description{
Search Zotero DB collection by name
}
