% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{title}
\alias{title}
\title{Title keyword search in Zotero DB}
\usage{
title(keyword, con = NULL, zotero_db = "~/Zotero/zotero.sqlite")
}
\arguments{
\item{keyword}{Title keyword (partial match)}

\item{con}{An existing DBI connection to a Zotero SQLite DB (optional)}

\item{zotero_db}{Path to the Zotero SQLite DB (default: '~/Zotero/zotero.sqlite')}
}
\value{
A data.frame with columns: itemID, title, creators, collections, itemTypeID, dateAdded, dateModified, key
}
\description{
Title keyword search in Zotero DB
}
