% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write_bib}
\alias{write_bib}
\title{Write a BIB export file}
\usage{
write_bib(bibs, path)
}
\arguments{
\item{bibs}{A \code{bibentry} object containing the bibliography entries.}

\item{path}{The file path where the BibTeX entries will be saved.}
}
\value{
None. The function writes the output directly to the specified file.
}
\description{
Write a BibTeX file with additional fields used by Web of Science/Clarivate for import.
}
